feat: socket join room
